requirements.txt
streamlit
tensorflow
tensorflow-hub
pandas
numpy
pillow
geopy
